people-and-planet-ai/weather-forecasting/requirements.txt
apache-beam[gcp]==2.41.0
earthengine-api==0.1.320
google-cloud-aiplatform==1.16.1
pandas==1.4.3
